relocate glu functions
